Upgrade Mapzen Android SDK (#822)
* gradle distribution 3.3, gradle version 2.3.3

* mapzen android sdk 1.4.0

* fix tests

* update test-results dir

+++ app/src/main/kotlin/com/mapzen/erasermap/model/ValhallaRouteManager.kt
@@ -54,7 +54,7 @@ class ValhallaRouteManager(val settings: AppSettings,
             val router = getInitializedRouter(type)
 
             if (location.hasBearing()) {
-                router.setLocation(start, location.bearing)
+                router.setLocation(start, location.bearing.toInt())
             } else {
                 router.setLocation(start)
             }
